+++ src/attributes/utils/handlers.ts
@@ -221,7 +221,9 @@ function getBooleanInput(attribute: AttributeInput) {
 function getDefaultInput(attribute: AttributeInput) {
   return {
     id: attribute.id,
-    values: attribute.value[0] === "" ? [] : attribute.value
+    values: ["", undefined, null].includes(attribute.value[0])
+      ? []
+      : attribute.value
   };
 }