+++ src/products/mutations.ts
@@ -330,7 +330,7 @@ export const variantUpdateMutation = gql`
     $addStocks: [StockInput!]!
     $removeStocks: [ID!]!
     $id: ID!
-    $attributes: [AttributeValueInput]
+    $attributes: [AttributeValueInput!]
     $sku: String
     $trackInventory: Boolean!
     $stocks: [StockInput!]!